# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ForfaitOperateur = typeof ForfaitOperateur[keyof typeof ForfaitOperateur];


export const ForfaitOperateur = {
  moov: 'moov',
  airtel: 'airtel',
} as const;

export interface Forfait {
  id: string;
  operateur: ForfaitOperateur;
  volume: string;
  prix: number;
  validite: string;
  validiteJours: number;
  description?: string;
}

export type PaiementInputPaymentOperator = typeof PaiementInputPaymentOperator[keyof typeof PaiementInputPaymentOperator];


export const PaiementInputPaymentOperator = {
  Airtel_Money: 'Airtel Money',
  Moov_Money: 'Moov Money',
} as const;

export interface PaiementInput {
  forfaitId: string;
  beneficiairePhone: string;
  paymentOperator: PaiementInputPaymentOperator;
  paymentPhone: string;
}

export interface OtpInput {
  reference: string;
  otp: string;
  forfaitId: string;
  beneficiairePhone: string;
  paymentOperator: string;
  paymentPhone: string;
  montant: number;
}

export interface PaiementReponse {
  transactionId: string;
  /** @nullable */
  reference?: string | null;
  statut: string;
  /** @nullable */
  montant?: number | null;
  /** @nullable */
  montantNet?: number | null;
  /** @nullable */
  operateur?: string | null;
  /** @nullable */
  telephone?: string | null;
}

export interface PaiementErreur {
  erreur: string;
  message: string;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  ussdCode?: string | null;
}

export interface StatutTransaction {
  transactionId: string;
  /** @nullable */
  reference?: string | null;
  statut: string;
  /** @nullable */
  montant?: number | null;
  /** @nullable */
  montantNet?: number | null;
  /** @nullable */
  telephone?: string | null;
  /** @nullable */
  confirmedAt?: string | null;
}

export interface Commande {
  id: number;
  forfaitId: string;
  operateurForfait: string;
  beneficiairePhone: string;
  paymentOperator: string;
  paymentPhone: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  reference?: string | null;
  statut: string;
  montant: number;
  createdAt: string;
}

export interface ApiError {
  error: string;
}