# lib/api-zod/src/generated/types/commande.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */

export interface Commande {
  id: number;
  forfaitId: string;
  operateurForfait: string;
  beneficiairePhone: string;
  paymentOperator: string;
  paymentPhone: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  reference?: string | null;
  statut: string;
  montant: number;
  createdAt: string;
}