# lib/api-zod/src/generated/types/forfait.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */
import type { ForfaitOperateur } from './forfaitOperateur';

export interface Forfait {
  id: string;
  operateur: ForfaitOperateur;
  volume: string;
  prix: number;
  validite: string;
  validiteJours: number;
  description?: string;
}