# lib/api-zod/src/generated/types/otpInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */

export interface OtpInput {
  reference: string;
  otp: string;
  forfaitId: string;
  beneficiairePhone: string;
  paymentOperator: string;
  paymentPhone: string;
  montant: number;
}