# lib/api-zod/src/generated/types/paiementErreur.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */

export interface PaiementErreur {
  erreur: string;
  message: string;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  ussdCode?: string | null;
}