# lib/api-zod/src/generated/types/paiementInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */
import type { PaiementInputPaymentOperator } from './paiementInputPaymentOperator';

export interface PaiementInput {
  forfaitId: string;
  beneficiairePhone: string;
  paymentOperator: PaiementInputPaymentOperator;
  paymentPhone: string;
}