# lib/api-zod/src/generated/types/paiementReponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Good Deal Tchad
 * OpenAPI spec version: 0.1.0
 */

export interface PaiementReponse {
  transactionId: string;
  /** @nullable */
  reference?: string | null;
  statut: string;
  /** @nullable */
  montant?: number | null;
  /** @nullable */
  montantNet?: number | null;
  /** @nullable */
  operateur?: string | null;
  /** @nullable */
  telephone?: string | null;
}